440: Split ballot paper section checks into separate files
