+++ apps/backend/src/controllers/ballotPaperSections/createBallotPaperSection.uc.ts
@@ -8,7 +8,7 @@ import {
 import type { Request, Response } from 'express';
 import { HttpStatusCode } from '../../httpStatusCode.js';
 import { createBallotPaperSection as createPersistentBallotPaperSection } from '../../services/ballotPaperSections.service.js';
-import { validateInsertableBallotPaperSection } from '../.bodyChecks/ballotPaperSectionChecks.js';
+import { validateInsertableBallotPaperSection } from '../.bodyChecks/ballotPaperSectionChecks/insertableBallotPaperSection.check.js';
 import { isBodyCheckValidationError } from '../.bodyChecks/bodyCheckValidationError.js';
 
 export const createBallotPaperSection = async (
