+++ apps/backend/src/controllers/ballotPaperSections/updateCandidateInBallotPaperSection.uc.ts
@@ -15,7 +15,7 @@ import {
   addCandidateToBallotPaperSection as addPersistentCandidateToBallotPaperSection,
   removeCandidateFromBallotPaperSection as removePersistentCandidateFromBallotPaperSection,
 } from '../../services/ballotPaperSections.service.js';
-import { validateUpdateableBallotPaperSectionCandidate } from '../.bodyChecks/ballotPaperSectionChecks.js';
+import { validateUpdateableBallotPaperSectionCandidate } from '../.bodyChecks/ballotPaperSectionChecks/updatableBallotPaperSectionCandidate.check.js';
 import { isBodyCheckValidationError } from '../.bodyChecks/bodyCheckValidationError.js';
 
 export const updateCandidateInBallotPaperSection = async (
