+++ apps/backend/test/ballotPaperSections/updateCandidateInBallotPaperSection.test.ts
@@ -12,7 +12,7 @@ import request from 'supertest';
 import { beforeAll, describe, expect, it } from 'vitest';
 import { app } from '../../src/app.js';
 import { generateUserTokens } from '../../src/auth/utils.js';
-import { BallotPaperCandidateValidationErrorMessage } from '../../src/controllers/.bodyChecks/ballotPaperSectionChecks.js';
+import { BallotPaperCandidateValidationErrorMessage } from '../../src/controllers/.bodyChecks/ballotPaperSectionChecks/common.check.js';
 import { HttpStatusCode } from '../../src/httpStatusCode.js';
 import { createUser, findUserBy } from '../../src/services/users.service.js';
 import {
